test(node): improve tests related to customized output filenames
