--- crates/rolldown/tests/rolldown/function/asset_filenames/entry.js
@@ -1,1 +1,0 @@
-import './main.css'
