--- crates/rolldown/tests/rolldown/function/entry_filenames/should_generate_specified_hash_length/_config.json
@@ -1,13 +1,0 @@
-{
-  "config": {
-    "input": [
-      {
-        "name": "entry",
-        "import": "./entry.js"
-      }
-    ],
-    "entryFilenames": "./entries/[hash]-[hash:11]-[hash:2]-[hash:22].js",
-    "chunkFilenames": "./chunks/[hash]-[hash:11]-[hash:2]-[hash:22].js"
-  },
-  "expectExecuted": false
-}
