--- crates/rolldown/tests/rolldown/function/entry_filenames/should_generate_specified_hash_length/entry.js
@@ -1,1 +1,0 @@
-export { value } from './shared.js'
