--- crates/rolldown/tests/rolldown/function/entry_filenames/should_generate_specified_hash_length/shared.js
@@ -1,1 +1,0 @@
-export const value = 'shared'
