--- packages/rolldown/tests/fixtures/output/hashFileNames/multi/main.css
@@ -1,4 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-}
