--- packages/rolldown/tests/fixtures/output/hashFileNames/multi/main.js
@@ -1,5 +1,0 @@
-import './main.css'
-
-console.log('1')
-const t = await import('./test.js')
-console.log(t)
