--- packages/rolldown/tests/fixtures/output/hashFileNames/multi/test.js
@@ -1,3 +1,0 @@
-import './test.css'
-
-export const hello = 'world'
