--- packages/rolldown/tests/fixtures/output/hashFileNames/normal/dynamic.js
@@ -1,1 +1,0 @@
-console.log('dynamic')
