--- packages/rolldown/tests/fixtures/output/hashFileNames/normal/entry.js
@@ -1,3 +1,0 @@
-import { value } from './shared.js'
-
-console.log(value)
