--- packages/rolldown/tests/fixtures/output/hashFileNames/normal/main.js
@@ -1,5 +1,0 @@
-import { value } from './shared.js'
-
-console.log(value)
-
-import('./dynamic.js')
