--- packages/rolldown/tests/fixtures/output/hashFileNames/normal/shared.js
@@ -1,1 +1,0 @@
-export const value = 1
